Fix base range circle not being shown during building placement
